+++ pyabc2/__init__.py
@@ -6,7 +6,7 @@ __version__ = "0.1.0"
 
 from .key import Key
 from .note import Note
-from .parse import Tune, _load_abcjs_if_in_jupyter
+from .parse import Tune
 from .pitch import Pitch, PitchClass
 
 __all__ = (
@@ -16,5 +16,3 @@ __all__ = (
     "PitchClass",
     "Tune",
 )
-
-_load_abcjs_if_in_jupyter()