+++ pyabc2/key.py
@@ -5,7 +5,6 @@ Key (e.g., G, Em, Ador)
 # https://github.com/campagnola/pyabc/blob/4c22a70a0f40ff82f608ffc19a1ca51a153f8c24/pyabc.py#L94
 import re
 import warnings
-from typing import Dict, List, Optional, Tuple
 
 from .pitch import PitchClass
 
@@ -106,7 +105,7 @@ we get A minor.
 """
 
 
-def _scale_chromatic_values(mode: str) -> List[int]:
+def _scale_chromatic_values(mode: str) -> list[int]:
     """The 7 integer chromatic values that make of the scale."""
     mode = _validate_and_normalize_mode_name(mode)
     i = MODE_SCALE_DEGREE[mode]
@@ -118,7 +117,7 @@ def _scale_chromatic_values(mode: str) -> List[int]:
     return [(v + dv) % 12 for v in vs_wrt_major]
 
 
-def _mode_chromatic_scale_degrees(mode: str, *, acc_fmt: str = "#") -> List[str]:
+def _mode_chromatic_scale_degrees(mode: str, *, acc_fmt: str = "#") -> list[str]:
     """For a given mode, scale degree representations for all 12 chromatic values."""
     valid_acc_formats = ["#", "b", "#/b", "b/#"]
     if acc_fmt not in valid_acc_formats:
@@ -158,7 +157,7 @@ def _mode_chromatic_scale_degrees(mode: str, *, acc_fmt: str = "#") -> List[str]
     return csds
 
 
-def _mode_scale_degrees_wrt_major(mode: str) -> List[str]:
+def _mode_scale_degrees_wrt_major(mode: str) -> list[str]:
     """ASCII scale degrees for a certain mode,
     referenced to major using #/b.
     """
@@ -172,10 +171,10 @@ def _mode_scale_degrees_wrt_major(mode: str) -> List[str]:
 
 
 def _scale_intervals(
-    values: List[int],
+    values: list[int],
     *,
     include_upper: bool = True,
-) -> List[str]:
+) -> list[str]:
     """Return list of intervals ('W' or 'H').
 
     Parameters
@@ -268,10 +267,10 @@ class Key:
     # TODO: maybe should move name to a .from_name for consistency with Pitch(Class)
     def __init__(
         self,
-        name: Optional[str] = None,
+        name: str | None = None,
         *,
-        tonic: Optional[str] = None,
-        mode: Optional[str] = None,
+        tonic: str | None = None,
+        mode: str | None = None,
     ):
         self.tonic: PitchClass
         """The tonic of the key."""
@@ -298,7 +297,7 @@ class Key:
         return _MODE_ABBR_TO_FULL[self._mode].capitalize()
 
     @staticmethod
-    def parse_key(key: str) -> Tuple[PitchClass, str]:
+    def parse_key(key: str) -> tuple[PitchClass, str]:
         """Parse a key spec string (e.g., ``D``, ``Amin``)
         and return the tonic and mode (3-char abbreviation).
 
@@ -309,7 +308,7 @@ class Key:
             raise ValueError(f"Invalid key specification '{key}'")
         base, acc, mode, extra = m.groups()
         if extra != "":
-            warnings.warn(f"extra info {extra!r} in key spec {key!r} ignored")
+            warnings.warn(f"extra info {extra!r} in key spec {key!r} ignored", stacklevel=2)
 
         if acc is None:
             acc = ""
@@ -331,7 +330,7 @@ class Key:
         # TODO: probably should either return a Key or be private / maybe outside class
 
     @property
-    def key_signature(self) -> List[str]:  # TODO: maybe change name
+    def key_signature(self) -> list[str]:  # TODO: maybe change name
         """
         List of accidentals that should be displayed in the key
         signature for the given key description.
@@ -353,7 +352,7 @@ class Key:
         return sig
 
     @property
-    def accidentals(self) -> Dict[str, str]:
+    def accidentals(self) -> dict[str, str]:
         """A dictionary of accidentals in the key signature,
         mapping natural note names to the accidental applied.
         """
@@ -416,7 +415,7 @@ class Key:
         return self.relative("minor")
 
     @property
-    def _letters(self) -> List[str]:
+    def _letters(self) -> list[str]:
         """Letters (natural note names) of the scale.
         Only depends on tonic.
         """
@@ -424,7 +423,7 @@ class Key:
         return CMAJ_LETTERS[ir:] + CMAJ_LETTERS[:ir]
 
     @property
-    def scale(self) -> List[PitchClass]:
+    def scale(self) -> list[PitchClass]:
         """Notes (pitch classes) of the scale."""
         nats = self._letters
         notes = [n + self.accidentals.get(n, "") for n in nats]
@@ -434,7 +433,7 @@ class Key:
         print(" ".join(f"{str(pc):2}" for pc in self.scale))
 
     @property
-    def scale_degrees_wrt_major(self) -> List[str]:
+    def scale_degrees_wrt_major(self) -> list[str]:
         """Scale degrees of the mode's scale, with #/b
         as compared to the major scale with the same tonic.
         """
@@ -449,7 +448,7 @@ class Key:
         print(" ".join(f"{s:{wmax}}" for s in csds))
 
     @property
-    def scale_chromatic_values(self) -> List[int]:
+    def scale_chromatic_values(self) -> list[int]:
         """Integer chromatic values that make up the scale,
         relative to the tonic.
         Only depends on mode.
@@ -460,7 +459,7 @@ class Key:
         print(" ".join(f"{str(i):2}" for i in self.scale_chromatic_values))
 
     @property
-    def intervals(self) -> List[str]:
+    def intervals(self) -> list[str]:
         """List of the intervals that compose the scale.
         Only depends on mode."""
         return _scale_intervals(self.scale_chromatic_values)